projects/project_two/UserManager/requirements.txt
Django==4.2.8
requests==2.25.1
requests-unixsocket==0.2.0
sqlparse==0.4.4
urllib3==1.26.11
six==1.16.0
python-dotenv==1.0.1
python-dateutil==2.8.2
Werkzeug==2.1.1
sanic==19.6.0
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
django-allauth==0.63.3
djangorestframework==3.15.2
PyJWT==2.8.0